Convert the string parameters for Prize()

+++ CompetitionUI/CreatePrizeForm.cs
@@ -52,10 +52,10 @@ namespace CompetitionUI
             }
 
             decimal prizeAmount = 0;
-            int prizePercentage = 0;
+            double prizePercentage = 0;
 
             bool prizeAmountValid = decimal.TryParse(prizeAmountValue.Text, out prizeAmount);
-            bool prizePercentageValid = int.TryParse(prizePercentageValue.Text, out prizePercentage);
+            bool prizePercentageValid = double.TryParse(prizePercentageValue.Text, out prizePercentage);
 
             if (!prizeAmountValid || !prizePercentageValid)
             {
